backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pytesseract
Pillow
openai
setuptools==75.8.0
wheel==0.45.1
langchain
langchain_openai
python-multipart
passlib
PyJWT
langchain-community
langgraph
pypdf
faiss-cpu
PyMuPDF==1.23.26